env: moving files and installing Bulma
