+++ app/assets/js/controllers.json
