+++ app/assets/js/controllers/hello_controller.js
