+++ app/assets/styles/styles.scss
@@ -1,1 +1,2 @@
-@import 'vendor/modularscale/modularscale';
+@import "vendor/modularscale/modularscale";
+@import "~bulma/bulma";
