+++ app/package.json
@@ -16,6 +16,7 @@
         "build": "encore production --progress"
     },
     "dependencies": {
+        "bulma": "^0.9.3",
         "node-sass": "^6.0.1",
         "postcss": "^8.3.9",
         "sass-loader": "^12.2.0"
